fix(db): error when creating new empty database
This commit adds two new migrations: one to create the database tables for
tags, projects, and settings, and another to seed default settings if they
do not already exist in the database. The `synchronize` option in the
`AppDataSource` configuration is set to `false` to prevent automatic schema
changes and ensure migrations are used instead.